Nosql/requirements.txt
flask==2.3.3
folium==0.14.0
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
geopy==2.3.0
psycopg2-binary==2.9.7
redis==4.6.0
sqlparse==0.4.4
gunicorn==21.2.0


